requirements.txt
coverage==5.0.4
pytest==5.4.1
paho-mqtt==1.5.0
redis==3.5.3
